# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "voiceit API",
    "description": "Production-quality Qwen3-TTS voice cloning API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/profiles": {
      "get": {
        "summary": "List Profiles",
        "description": "List all voice profiles.",
        "operationId": "list_profiles_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VoiceProfileResponse"
                  },
                  "type": "array",
                  "title": "Response List Profiles Profiles Get"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "post": {
        "summary": "Create Profile",
        "description": "Create a new voice profile.",
        "operationId": "create_profile_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/profiles/{profile_id}": {
      "get": {
        "summary": "Get Profile",
        "description": "Get a voice profile by ID.",
        "operationId": "get_profile_profiles__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "put": {
        "summary": "Update Profile",
        "description": "Update a voice profile.",
        "operationId": "update_profile_profiles__profile_id__put",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceProfileCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "delete": {
        "summary": "Delete Profile",
        "description": "Delete a voice profile.",
        "operationId": "delete_profile_profiles__profile_id__delete",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/profiles/{profile_id}/samples": {
      "post": {
        "summary": "Add Profile Sample",
        "description": "Add a sample to a voice profile.",
        "operationId": "add_profile_sample_profiles__profile_id__samples_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_profile_sample_profiles__profile_id__samples_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSampleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "get": {
        "summary": "Get Profile Samples",
        "description": "Get all samples for a profile.",
        "operationId": "get_profile_samples_profiles__profile_id__samples_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileSampleResponse"
                  },
                  "title": "Response Get Profile Samples Profiles  Profile Id  Samples Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/profiles/samples/{sample_id}": {
      "delete": {
        "summary": "Delete Profile Sample",
        "description": "Delete a profile sample.",
        "operationId": "delete_profile_sample_profiles_samples__sample_id__delete",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_add_profile_sample_profiles__profile_id__samples_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "reference_text": {
            "type": "string",
            "title": "Reference Text"
          }
        },
        "type": "object",
        "required": [
          "file",
          "reference_text"
        ],
        "title": "Body_add_profile_sample_profiles__profile_id__samples_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileSampleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "audio_path": {
            "type": "string",
            "title": "Audio Path"
          },
          "reference_text": {
            "type": "string",
            "title": "Reference Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "profile_id",
          "audio_path",
          "reference_text"
        ],
        "title": "ProfileSampleResponse",
        "description": "Response model for profile sample."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceProfileCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "language": {
            "type": "string",
            "pattern": "^(en|zh)$",
            "title": "Language",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "VoiceProfileCreate",
        "description": "Request model for creating a voice profile."
      },
      "VoiceProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "language",
          "created_at",
          "updated_at"
        ],
        "title": "VoiceProfileResponse",
        "description": "Response model for voice profile."
      }
    }
  },
...
}